Comments, refactoring
