+++ Termoservis/Termoservis.Common/Extensions/StringExtensions.cs
@@ -1,6 +1,4 @@
-﻿using System.Globalization;
-
-namespace Termoservis.Common.Extensions
+﻿namespace Termoservis.Common.Extensions
 {
 	/// <summary>
 	/// The <see cref="string"/> extensions.
