+++ Termoservis/Termoservis.DAL/Repositories/ICustomerDevicesRepository.cs
@@ -3,6 +3,10 @@ using Termoservis.Models;
 
 namespace Termoservis.DAL.Repositories
 {
+    /// <summary>
+    /// The customer devices repository.
+    /// </summary>
+    /// <seealso cref="IAddRepository{CustomerDevice, Int64}" />
     public interface ICustomerDevicesRepository : IAddRepository<CustomerDevice, long>
     {
     }
