+++ Termoservis/Termoservis.DAL/Repositories/IWorkItemsRepository.cs
@@ -3,6 +3,10 @@ using Termoservis.Models;
 
 namespace Termoservis.DAL.Repositories
 {
+    /// <summary>
+    /// WorkItem repository.
+    /// </summary>
+    /// <seealso cref="IAddRepository{WorkItem, Int64}" />
     public interface IWorkItemsRepository : IAddRepository<WorkItem, long>
     {
     }
