+++ Termoservis/Termoservis.MigrationTool/Program.cs
@@ -7,7 +7,6 @@ using System.Linq;
 using System.Runtime.InteropServices;
 using System.Text;
 using System.Threading;
-using System.Threading.Tasks;
 using CsvHelper;
 using Termoservis.Common.Extensions;
 using Termoservis.DAL;
