+++ Termoservis/Termoservis.Models/Worker.cs
@@ -4,12 +4,27 @@ using System.ComponentModel.DataAnnotations.Schema;
 
 namespace Termoservis.Models
 {
+    /// <summary>
+    /// The worker model.
+    /// </summary>
     public class Worker
     {
+        /// <summary>
+        /// Gets or sets the identifier.
+        /// </summary>
+        /// <value>
+        /// The identifier.
+        /// </value>
         [Key]
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
         public long Id { get; set; }
 
+        /// <summary>
+        /// Gets or sets the name.
+        /// </summary>
+        /// <value>
+        /// The name.
+        /// </value>
         [Required]
         [DisplayName("Ime i prezime")]
         public string Name { get; set; }
