+++ Termoservis/Termoservis.Web/App_Start/UnityConfig.cs
@@ -9,7 +9,7 @@ using Termoservis.DAL;
 using Termoservis.DAL.Repositories;
 using Termoservis.Models;
 using Termoservis.Web.Services;
-using Termoservis.Web.Controllers;
+using Termoservis.BLL;
 
 namespace Termoservis.Web
 {
@@ -57,6 +57,8 @@ namespace Termoservis.Web
 	        container.RegisterType<IAddressesRepository, AddressesRepository>();
 			container.RegisterType<ITelephoneNumbersRepository, TelephoneNumbersRepository>();
 			container.RegisterType<ICustomersRepository, CustomersRepository>();
+            container.RegisterType<ICustomerDevicesRepository, CustomerDevicesRepository>();
+            container.RegisterType<IWorkItemsRepository, WorkItemsRepository>();
 
             // Register services
             container.RegisterType<ICustomerService, CustomerService>();
