+++ Termoservis/Termoservis.Web/Controllers/WorkItemController.cs
@@ -2,6 +2,10 @@ using System.Web.Mvc;
 
 namespace Termoservis.Web.Controllers
 {
+    /// <summary>
+    /// The work item controller.
+    /// </summary>
+    /// <seealso cref="Controller" />
     [Authorize]
     [RequireHttps]
     public class WorkItemController : Controller
