+++ Termoservis/Termoservis.Web/Helpers/AjaxAttribute.cs
@@ -15,12 +15,24 @@ namespace Termoservis.Web.Helpers
         private readonly bool ajax;
 
 
+        /// <summary>
+        /// Initializes a new instance of the <see cref="AjaxAttribute"/> class.
+        /// </summary>
+        /// <param name="ajax">if set to <c>true</c> action is Ajax action.</param>
         public AjaxAttribute(bool ajax)
         {
             this.ajax = ajax;
         }
 
 
+        /// <summary>
+        /// Determines whether the action method selection is valid for the specified controller context.
+        /// </summary>
+        /// <param name="controllerContext">The controller context.</param>
+        /// <param name="methodInfo">Information about the action method.</param>
+        /// <returns>
+        /// true if the action method selection is valid for the specified controller context; otherwise, false.
+        /// </returns>
         public override bool IsValidForRequest(ControllerContext controllerContext, MethodInfo methodInfo)
         {
             return ajax == controllerContext.HttpContext.Request.IsAjaxRequest();
