+++ Termoservis/Termoservis.Web/Models/Customer/CustomerFormViewModel.cs
@@ -12,16 +12,29 @@ namespace Termoservis.Web.Models.Customer
 	/// <seealso cref="ILocationViewModel" />
 	public class CustomerFormViewModel : Termoservis.Models.Customer, ILocationViewModel, IFormViewModel
     {
-	    public CustomerFormViewModel()
+        /// <summary>
+        /// Initializes a new instance of the <see cref="CustomerFormViewModel"/> class.
+        /// </summary>
+        public CustomerFormViewModel()
 	    {
 	    }
 
-	    public CustomerFormViewModel(string actionName)
+        /// <summary>
+        /// Initializes a new instance of the <see cref="CustomerFormViewModel"/> class.
+        /// </summary>
+        /// <param name="actionName">Name of the action.</param>
+        public CustomerFormViewModel(string actionName)
 	    {
 	        this.ActionName = actionName;
 	    }
 
 
+        /// <summary>
+        /// Gets or sets the name of the action.
+        /// </summary>
+        /// <value>
+        /// The name of the action.
+        /// </value>
         public string ActionName { get; set; }
 
 		/// <summary>
