+++ Termoservis/Termoservis.Web/Models/Customer/IFormViewModel.cs
@@ -1,7 +1,16 @@
 ﻿namespace Termoservis.Web.Models.Customer
 {
+    /// <summary>
+    /// The form view model.
+    /// </summary>
     public interface IFormViewModel
     {
+        /// <summary>
+        /// Gets or sets the name of the action.
+        /// </summary>
+        /// <value>
+        /// The name of the action.
+        /// </value>
         string ActionName { get; set; }
     }
 }
