+++ Termoservis/Termoservis.Web/Views/Customers/Index.cshtml
@@ -1,5 +1,4 @@
-﻿@using Termoservis.Web.Helpers
-@model Termoservis.Web.Controllers.CustomersSearchResult
+﻿@model Termoservis.Web.Models.Customer.CustomersSearchResult
 
 @{
     ViewBag.Title = "Korisnici";
