+++ Termoservis/Termoservis.Web/Views/Customers/_CustomersTablePartial.cshtml
@@ -1,5 +1,5 @@
 ﻿@using Termoservis.Web.Helpers
-@model Termoservis.Web.Controllers.CustomersSearchResult
+@model Termoservis.Web.Models.Customer.CustomersSearchResult
 
         @using (Html.BeginForm())
         {
