+++ Termoservis/Termoservis.Web/Views/Customers/_WorkItemDetailsPartial.cshtml
@@ -1,5 +1,4 @@
-﻿@using Termoservis.Models
-@model Termoservis.Web.Models.Customer.WorkItemFormViewModel
+﻿@model Termoservis.Web.Models.Customer.WorkItemFormViewModel
 
 @using (Html.BeginForm(Model.ActionName, "WorkItem", FormMethod.Post, new {@class = "col-xs-12", role = "form", style = "background-color: rgb(249, 249, 249)" }))
 {
